扩展合成地址支持的最大公钥数量；解决内存池清理问题；修复写入区块链数据库时的一个错误；

+++ AntSharesCore/Core/Block.cs
@@ -260,7 +260,7 @@ namespace AntShares.Core
                 }
             }
             ECCPublicKey[] pubkeys = miners.OrderByDescending(p => p.Value).Select(p => p.Key).Concat(Blockchain.StandbyMiners).Take(miner_count).ToArray();
-            if (Miner != Wallet.CreateRedeemScript((byte)Blockchain.GetMinSignatureCount(miner_count), pubkeys).ToScriptHash())
+            if (Miner != Wallet.CreateRedeemScript(Blockchain.GetMinSignatureCount(miner_count), pubkeys).ToScriptHash())
                 result |= VerificationResult.InvalidSignature;
             if (completely)
             {
